requirements.txt
Django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.1
ruff==0.7.3
aws-cdk-lib==2.166.0
constructs>=10.0.0,<11.0.0